dosfs: Unsigned compared against 0
Fix the type of "cmpltd" to be ssize_t so that assigning it to -1
will terminate processing as intended.

+++ cpukit/libfs/src/dosfs/msdos_dir.c
@@ -78,7 +78,7 @@ msdos_dir_read(rtems_libio_t *iop, void *buffer, size_t count)
     const size_t       buf_size = converter->buffer.size;
     uint32_t           start = 0;
     ssize_t            ret = 0;
-    uint32_t           cmpltd = 0;
+    ssize_t            cmpltd = 0;
     uint32_t           j = 0, i = 0;
     uint32_t           bts2rd = 0;
     uint32_t           cur_cln = 0;
